+++ saleor/graphql/account/schema.py
@@ -6,13 +6,11 @@ from ..core.types import FilterInputObjectType
 from ..decorators import one_of_permissions_required, permission_required
 from .bulk_mutations import CustomerBulkDelete, StaffBulkDelete, UserBulkSetActive
 from .deprecated.mutations_service_account import (
-    ServiceAccountClearPrivateMeta,
     ServiceAccountCreate,
     ServiceAccountDelete,
     ServiceAccountTokenCreate,
     ServiceAccountTokenDelete,
     ServiceAccountUpdate,
-    ServiceAccountUpdatePrivateMeta,
 )
 from .deprecated.resolvers import resolve_service_accounts
 from .deprecated.sorters import ServiceAccountSortingInput
@@ -28,7 +26,6 @@ from .mutations.account import (
     AccountRequestDeletion,
     AccountSetDefaultAddress,
     AccountUpdate,
-    AccountUpdateMeta,
     ConfirmEmailChange,
     RequestEmailChange,
 )
@@ -37,8 +34,6 @@ from .mutations.base import (
     PasswordChange,
     RequestPasswordReset,
     SetPassword,
-    UserClearMeta,
-    UserUpdateMeta,
 )
 from .mutations.jwt import (
     CreateToken,
@@ -64,8 +59,6 @@ from .mutations.staff import (
     StaffUpdate,
     UserAvatarDelete,
     UserAvatarUpdate,
-    UserClearPrivateMeta,
-    UserUpdatePrivateMeta,
 )
 from .resolvers import (
     resolve_address,
@@ -248,13 +241,6 @@ class AccountMutations(graphene.ObjectType):
     account_request_deletion = AccountRequestDeletion.Field()
     account_delete = AccountDelete.Field()
 
-    account_update_meta = AccountUpdateMeta.Field(
-        deprecation_reason=(
-            "Use the `updateMetadata` mutation. This field will be removed after "
-            "2020-07-31."
-        )
-    )
-
     # Staff mutations
     address_create = AddressCreate.Field()
     address_update = AddressUpdate.Field()
@@ -275,32 +261,6 @@ class AccountMutations(graphene.ObjectType):
     user_avatar_delete = UserAvatarDelete.Field()
     user_bulk_set_active = UserBulkSetActive.Field()
 
-    user_update_metadata = UserUpdateMeta.Field(
-        deprecation_reason=(
-            "Use the `updateMetadata` mutation. This field will be removed after "
-            "2020-07-31."
-        )
-    )
-    user_clear_metadata = UserClearMeta.Field(
-        deprecation_reason=(
-            "Use the `deleteMetadata` mutation. This field will be removed after "
-            "2020-07-31."
-        )
-    )
-
-    user_update_private_metadata = UserUpdatePrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `updatePrivateMetadata` mutation. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-    user_clear_private_metadata = UserClearPrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `deletePrivateMetadata` mutation. This field will be removed "
-            "after 2020-07-31."
-        )
-    )
-
     service_account_create = ServiceAccountCreate.Field(
         deprecation_reason=(
             "Use the `appCreate` mutation instead. This field will be removed after "
@@ -319,20 +279,6 @@ class AccountMutations(graphene.ObjectType):
             "2020-07-31."
         )
     )
-
-    service_account_update_private_metadata = ServiceAccountUpdatePrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `updatePrivateMetadata` mutation with App instead."
-            "This field will be removed after 2020-07-31."
-        )
-    )
-    service_account_clear_private_metadata = ServiceAccountClearPrivateMeta.Field(
-        deprecation_reason=(
-            "Use the `deletePrivateMetadata` mutation with App instead."
-            "This field will be removed after 2020-07-31."
-        )
-    )
-
     service_account_token_create = ServiceAccountTokenCreate.Field(
         deprecation_reason=(
             "Use the `appTokenCreate` mutation instead. This field will be removed "